+++ src/type_pool.cpp
@@ -65,7 +65,7 @@ TypePool::TypePool() {
   this->initBuiltinType(TYPE::StringIter, "StringIter%%", false, TYPE::Any, info_StringIterType());
   this->initBuiltinType(TYPE::UnixFD, "FD", false, TYPE::Any, info_UnixFDType());
   this->initBuiltinType(TYPE::Reader, "Reader%", false, TYPE::Any, info_ReaderType());
-  this->initBuiltinType(TYPE::Command, "Command", false, TYPE::Any, info_Dummy());
+  this->initBuiltinType(TYPE::Command, "Command", false, TYPE::Any, info_CommandType());
 
   // initialize type template
   std::vector<const DSType *> elements = {&this->get(TYPE::Any)};